refactor(sync-client): split DesktopSyncEngineTest into focused test classes
